Move http server and client, get rid of ssl there
