+++ src/Rpc/RpcServer.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "HttpServer.h"
+#include "HTTP/HttpServer.h"
 
 #include <functional>
 #include <unordered_map>
